Create separate file for mock variables
