+++ src/constants/mockVariables.ts
@@ -1,0 +1,22 @@
+import { Dream } from "../models/types.js";
+
+export const dreams: Dream[] = [
+  {
+    id: 1,
+    name: "Lära mig HTML/CSS",
+    theme: "teknikdrömmar",
+    isChecked: true,
+  },
+  {
+    id: 2,
+    name: "Lära mig TypeScript",
+    theme: "teknikdrömmar",
+    isChecked: false,
+  },
+  {
+    id: 3,
+    name: "En dröm som tar flera rader lorem ipsum",
+    theme: "vardagsdrömmar",
+    isChecked: false,
+  },
+];
